refactor: user을 auth 모듈로 분리
