+++ src/auth/auth.module.ts
@@ -1,0 +1,17 @@
+import { Logger, Module } from '@nestjs/common';
+import { AuthService } from './auth.service';
+import { AuthController } from './auth.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { UserEntity } from 'src/users/entities/user.entity';
+import { PassportModule } from '@nestjs/passport';
+import { JwtModule } from '@nestjs/jwt';
+import { GoogleStrategy } from './strategies/google.strategy';
+import { JwtStrategy } from './strategies/jwt.access.strategy';
+import { JwtRefreshStrategy } from './strategies/jwt.refresh.strategy';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([UserEntity]), PassportModule, JwtModule.register({})],
+  controllers: [AuthController],
+  providers: [AuthService, Logger, GoogleStrategy, JwtStrategy, JwtRefreshStrategy],
+})
+export class AuthModule {}
