+++ src/configs/swagger.config.ts
@@ -1,0 +1,9 @@
+import { DocumentBuilder } from '@nestjs/swagger';
+
+export class SwaggerConfig {
+  public builder = new DocumentBuilder();
+
+  public initializeOptions() {
+    return this.builder.setTitle('OHA').setVersion('1.0').build();
+  }
+}
