--- src/users/dto/create-user.dto.ts
@@ -1,1 +1,0 @@
-export class CreateUserDto {}
